added all file operation, almost all dir operation, formatting and etc.

+++ FileSystemNTFS/Controller/FileSystemController/DeleteFile.cs
@@ -17,9 +17,16 @@ namespace FileSystemNTFS.BL.Controller.FileSystemController
                 return false;
             }
 
+            var mftItem = FileSystem.MFTController.MFT.Entries.SingleOrDefault(x => x.Attributes.FullPath.Equals(fullPath, StringComparison.OrdinalIgnoreCase));
+            if(mftItem == null)
+               throw new ArgumentNullException("Такой записи не существует!",nameof(mftItem));
+
+            FileSystem.SuperblockController.FreeAllClustersMFTEntry(mftItem);
+            DeleteMFTDataFromDir(mftItem);
+            FileSystem.MFTController.MFT.Entries.Remove(mftItem);
             File.Delete(fullPath);
 
-
+            Save();
 
             return true;
         }
